Refactor student selection process
